backend/requirements.txt
python-dotenv
google-generativeai
pytest
pytest-asyncio
openai-agents
litellm